pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rajoybot"
version = "1.0.0"
description = "Telegram bot that serves audio clips of Mariano Rajoy quotes"
requires-python = ">=3.12"
license = "GPL-3.0-only"
dependencies = [
    "python-telegram-bot~=22.7",
    "unidecode~=1.4",
    "tortoise-orm~=1.1",
    "aiosqlite~=0.22",
    "aiomysql~=0.3",
    "python-dotenv~=1.2",
]

[project.optional-dependencies]
dev = [
    "ruff~=0.11",
    "mypy~=1.15",
    "pytest~=8.3",
    "pytest-cov~=6.1",
    "pytest-asyncio~=0.25",
]

[tool.setuptools]
packages = []

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.12"
check_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
